feature: init flutterfire
